+++ question-service/src/main/java/com/akash/questionservice/dao/QuestionDao.java
@@ -14,5 +14,5 @@ public interface QuestionDao extends JpaRepository<Question, Long> {
     List<Question> findByCategory(String category);
 
     @Query(value = "SELECT q.id FROM question q WHERE q.category=:category ORDER BY RANDOM() LIMIT :numQ", nativeQuery = true)
-    List<Integer> findRandomQuestions(String category, Integer numQ);
+    List<Long> findRandomQuestions(String category, Integer numQ);
 }
